+++ src/SlideDotNet/Models/PresentationEx.cs
@@ -17,28 +17,19 @@ namespace SlideDotNet.Models
     {
         #region Fields
 
-        private PresentationDocument _xmlDoc;
-        private readonly Lazy<ISlideCollection> _slides;
+        private PresentationDocument _sdkPre;
+        private readonly Lazy<EditAbleCollection<Slide>> _slides;
         private bool _disposed;
 
         #endregion Fields
 
         #region Properties
 
-        /// <summary>
-        /// <inheritdoc cref="IPresentation.Slides"/>
-        /// </summary>
-        public ISlideCollection Slides => _slides.Value;
+        public EditAbleCollection<Slide> Slides => _slides.Value;
 
-        /// <summary>
-        /// <inheritdoc cref="IPresentation.SlideWidth"/>
-        /// </summary>
-        public int SlideWidth => _xmlDoc.PresentationPart.Presentation.SlideSize.Cx.Value;
+        public int SlideWidth => _sdkPre.PresentationPart.Presentation.SlideSize.Cx.Value;
 
-        /// <summary>
-        /// <inheritdoc cref="IPresentation.SlideHeight"/>
-        /// </summary>
-        public int SlideHeight => _xmlDoc.PresentationPart.Presentation.SlideSize.Cy.Value;
+        public int SlideHeight => _sdkPre.PresentationPart.Presentation.SlideSize.Cy.Value;
 
         #endregion Properties
 
@@ -50,9 +41,9 @@ namespace SlideDotNet.Models
         public PresentationEx(string pptxPath)
         {
             ThrowIfInvalid(pptxPath);
-            _xmlDoc = PresentationDocument.Open(pptxPath, true);
+            _sdkPre = PresentationDocument.Open(pptxPath, true);
             ThrowIfSlidesNumberLarge();
-            _slides = new Lazy<ISlideCollection>(InitSlides);
+            _slides = new Lazy<EditAbleCollection<Slide>>(InitSlides);
         }
 
         /// <summary>
@@ -62,9 +53,9 @@ namespace SlideDotNet.Models
         public PresentationEx(Stream pptxStream)
         {
             ThrowIfInvalid(pptxStream);
-            _xmlDoc = PresentationDocument.Open(pptxStream, true);
+            _sdkPre = PresentationDocument.Open(pptxStream, true);
             ThrowIfSlidesNumberLarge();
-            _slides = new Lazy<ISlideCollection>(InitSlides);
+            _slides = new Lazy<EditAbleCollection<Slide>>(InitSlides);
         }
 
         /// <summary>
@@ -76,50 +67,39 @@ namespace SlideDotNet.Models
             ThrowIfInvalid(pptxBytes);
             var pptxStream = new MemoryStream();
             pptxStream.Write(pptxBytes, 0, pptxBytes.Length);
-            _xmlDoc = PresentationDocument.Open(pptxStream, true);
+            _sdkPre = PresentationDocument.Open(pptxStream, true);
             ThrowIfSlidesNumberLarge();
-            _slides = new Lazy<ISlideCollection>(InitSlides);
+            _slides = new Lazy<EditAbleCollection<Slide>>(InitSlides);
         }
 
         #endregion Constructors
 
         #region Public Methods
 
-        /// <summary>
-        /// <inheritdoc cref="IPresentation.SaveAs"/>
-        /// </summary>
         public void SaveAs(string filePath)
         {
             Check.NotEmpty(filePath, nameof(filePath));
-            _xmlDoc = (PresentationDocument)_xmlDoc.SaveAs(filePath);
+            _sdkPre = (PresentationDocument)_sdkPre.SaveAs(filePath);
         }
 
-        /// <summary>
-        /// <inheritdoc cref="IPresentation.SaveAs"/> //TODO: resolve inheritdoc conflicts
-        /// </summary>
-        /// <param name="stream"></param>
         public void SaveAs(Stream stream)
         {
             Check.NotNull(stream, nameof(stream));
-            _xmlDoc = (PresentationDocument)_xmlDoc.Clone(stream);
+            _sdkPre = (PresentationDocument)_sdkPre.Clone(stream);
         }
 
-        /// <summary>
-        /// <inheritdoc cref="IPresentation.Close"/>
-        /// </summary>
         public void Close()
         {
             if (_disposed)
             {
                 return;
             }
-            _xmlDoc.Close();
+            _sdkPre.Close();
             _disposed = true;
+
+            
         }
 
-        /// <summary>
-        /// Saves and closes the current presentation.
-        /// </summary>
         public void Dispose()
         {
             Close();
@@ -129,10 +109,10 @@ namespace SlideDotNet.Models
 
         #region Private Methods
 
-        private ISlideCollection InitSlides()
+        private EditAbleCollection<Slide> InitSlides()
         {
-            var preSettings = new PreSettings(_xmlDoc.PresentationPart.Presentation);
-            var slideCollection = SlideCollection.Create(_xmlDoc, preSettings);
+            var preSettings = new PreSettings(_sdkPre.PresentationPart.Presentation);
+            var slideCollection = SlideCollection.Create(_sdkPre, preSettings);
 
             return slideCollection;
         }
@@ -172,7 +152,7 @@ namespace SlideDotNet.Models
 
         private void ThrowIfSlidesNumberLarge()
         {
-            var nbSlides = _xmlDoc.PresentationPart.SlideParts.Count();
+            var nbSlides = _sdkPre.PresentationPart.SlideParts.Count();
             if (nbSlides > Limitations.MaxSlidesNumber)
             {
                 Close();