+++ src/SlideDotNet/Services/ShapeFactory.cs
@@ -5,6 +5,7 @@ using DocumentFormat.OpenXml.Packaging;
 using SlideDotNet.Models.Settings;
 using SlideDotNet.Models.SlideComponents;
 using SlideDotNet.Services.Placeholders;
+using SlideDotNet.Services.ShapeCreators;
 using P = DocumentFormat.OpenXml.Presentation;
 using A = DocumentFormat.OpenXml.Drawing;
 // ReSharper disable PossibleMultipleEnumeration
@@ -56,7 +57,7 @@ namespace SlideDotNet.Services
 
             sdkShapeHandler.Successor = sdkGroupShapeHandler;
             sdkGroupShapeHandler.Successor = oleGrFrameHandler;
-            // OLE objects handler must be before pictures handler, cause OLE container can contain p:pic elements, thereby OLE as a picture
+            // OLE objects handler must be before pictures handler, cause OLE container can contain p:pic elements, thereby defining OLE as a picture
             oleGrFrameHandler.Successor = pictureHandler;
             pictureHandler.Successor = chartGrFrameHandler;
             chartGrFrameHandler.Successor = tableGrFrameHandler;